+++ libs/profiling/tests/test_cli.py
@@ -3,7 +3,6 @@ from typing import List
 
 import pytest
 from click.testing import CliRunner
-
 from vivarium.profiling.tools.cli import profile_sim